+++ Source/Modules/Catalog/Gapotchenko.FX.Text/StringPolyfills.cs
@@ -4,8 +4,6 @@
 
 namespace Gapotchenko.FX.Text;
 
-#pragma warning disable CA1062
-
 /// <summary>
 /// Provides polyfill methods for <see cref="string"/> type.
 /// </summary>